pyproject.toml
[project]
name = "sixboxweb"
version = "0.1.0"
description = "web"
authors = [
    {name = "skily"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "tinydb (>=4.8.2,<5.0.0)",  # 数据库
    "flask (>=3.1.2,<4.0.0)",   # 核心
    "pyjwt (>=2.10.1,<3.0.0)",  # 用于token生成/解析
    "babel (>=2.17.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
